refactor: HttpRequest 클래스의 패키지 변경
